Add ImageDataSource and ImageView classes
The ImageDataSource provides an interface that can be used either in the
live preview (where chunks of the image are streamed through a
websocket) or in the post-processing, where full images are fetched in
rest calls to the server.

ImageView gets the data from the ImageDataSource, and draws it into an
actual image after colormapping the raw data. In the future additional
controls such as selecting pixels or drawing will be added.